moved ElectionDetails to ElectionShow
