+++ web/frontend/src/App.tsx
@@ -18,6 +18,7 @@ import About from './pages/About';
 import Admin from './pages/Admin';
 import ElectionIndex from './pages/ElectionIndex';
 import ElectionCreate from './pages/ElectionCreate';
+import ElectionShow from './pages/ElectionShow';
 import ResultIndex from './pages/ResultIndex';
 import ResultShow from './pages/ResultShow';
 import BallotIndex from './pages/BallotIndex';
@@ -26,7 +27,6 @@ import NavBar from './pages/NavBar';
 import Footer from './pages/Footer';
 
 import { GET_PERSONNAL_INFOS } from './components/utils/ExpressEndoints';
-import ElectionDetails from './components/ElectionDetails';
 import './App.css';
 
 const App: FC = () => {
@@ -63,7 +63,7 @@ const App: FC = () => {
                 <Route path={ROUTE_HOME} element={<Home isLogged={isLogged} />} />
                 <Route path={ROUTE_ELECTION_INDEX} element={<ElectionIndex />} />
                 <Route path={ROUTE_ELECTION_CREATE} element={<ElectionCreate />} />
-                <Route path={ROUTE_ELECTION_SHOW + '/:electionId'} element={<ElectionDetails />} />
+                <Route path={ROUTE_ELECTION_SHOW + '/:electionId'} element={<ElectionShow />} />
                 <Route path={ROUTE_RESULT_INDEX} element={<ResultIndex />} />
                 <Route path={ROUTE_RESULT_SHOW + '/:electionId'} element={<ResultShow />} />
                 <Route path={ROUTE_BALLOT_INDEX} element={<BallotIndex />} />
