+++ web/frontend/src/pages/ElectionShow.css
